Align ScaleDropdown better with fluent base
