+++ packages/fluent/src/components/ScaleDropdown/ScaleDropdown.hooks.ts
@@ -2,21 +2,30 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import type { IDropdownOption } from '@fluentui/react'
+import type {
+	IDropdownOption,
+	IDropdownStyleProps,
+	IDropdownStyles,
+	IStyleFunctionOrObject,
+} from '@fluentui/react'
 import type {
 	ContinuousColorScaleFunction,
 	NominalColorScaleFunction,
 } from '@thematic/core'
 import { chooseScale } from '@thematic/core'
 import { useThematic } from '@thematic/react'
+import { useSize } from 'ahooks'
+import merge from 'lodash-es/merge.js'
 import type React from 'react'
 import { useMemo } from 'react'
 
-import type { ChipsProps } from '../ScaleDropdown.types.js'
-import { selectColorPalette } from '../ScaleDropdown.utils.js'
+import type { ChipsProps } from './ScaleDropdown.types.js'
+import { selectColorPalette } from './ScaleDropdown.utils.js'
+
+const DEFAULT_WIDTH = 200
+const DEFAULT_HEIGHT = 32
 
 const ITEM_LEFT_PADDING = 8 // default right padding in fluent item
-const ITEM_BORDER_MODIFIER = 1 // accounts for transparent border on outer container
 const CARET_PADDING = 28 // defined default in fluent dropdown is 28 (this also aligns with item right padding)
 export const TEXT_WIDTH = 80 // TODO: adjust this based on font size/max measured
 const LABEL_HEIGHT = 29 // defined default in fluent dropdown
@@ -25,35 +34,32 @@ const LABEL_HEIGHT = 29 // defined default in fluent dropdown
 // visually we'll keep it lowercase in this app for visual consistency
 const TITLE_CASE = false
 
+export function useDropdownStyles(
+	styles?: IStyleFunctionOrObject<IDropdownStyleProps, IDropdownStyles>,
+): IStyleFunctionOrObject<IDropdownStyleProps, IDropdownStyles> {
+	return useMemo(
+		() =>
+			merge(
+				{
+					root: {
+						textAlign: 'left',
+					},
+				},
+				styles,
+			),
+		[styles],
+	)
+}
+
 export function usePaletteWidth(width: number): number {
 	// subtract space for the caret, pad, text, etc.
-	return (
-		width -
-		ITEM_BORDER_MODIFIER -
-		ITEM_LEFT_PADDING -
-		TEXT_WIDTH -
-		CARET_PADDING -
-		ITEM_BORDER_MODIFIER
-	)
+	return width - ITEM_LEFT_PADDING - TEXT_WIDTH - CARET_PADDING
 }
 
 export function usePaletteHeight(height: number, label?: string): number {
 	// the measured component dimensions will include the label if present
 	const root = label ? height - LABEL_HEIGHT : height
 	return root / 2
-}
-
-/**
- * Provides style overrides for root dropdown container
- * @returns
- */
-export function useContainerStyle(): React.CSSProperties {
-	return useMemo(
-		() => ({
-			textAlign: 'left',
-		}),
-		[],
-	)
 }
 
 /**
@@ -66,7 +72,7 @@ export function useItemStyle(width: number): React.CSSProperties {
 	return useMemo(
 		() => ({
 			width: width - CARET_PADDING,
-			paddingLeft: ITEM_BORDER_MODIFIER,
+			paddingLeft: 0,
 			paddingRight: CARET_PADDING,
 		}),
 		[width],
@@ -109,4 +115,33 @@ export function useScale(
 
 export function usePaletteComponent(key: string): React.FC<ChipsProps> {
 	return useMemo(() => selectColorPalette(key), [key])
+}
+
+export interface Dimensions {
+	width: number
+	height: number
+}
+
+export function useSafeDimensions(
+	ref: React.RefObject<HTMLDivElement>,
+): Dimensions {
+	const dimensions = useSize(ref)
+	return (
+		dimensions || {
+			width: DEFAULT_WIDTH,
+			height: DEFAULT_HEIGHT,
+		}
+	)
+}
+
+/**
+ * Retrieve a non-zero width/height, because collapsible panels can force invalid color arrays
+ * @param width
+ * @param height
+ */
+export function useSafeCollapseDimensions(
+	width: number,
+	height: number,
+): [number, number] {
+	return [width <= 0 ? 1 : width, height <= 0 ? 1 : height]
 }
